refactor: Use query in stablecoin price change filename
