+++ apps/web/src/views/AddLiquidityV3/formViews/V3FormView/components/PositionPreview.tsx
@@ -5,7 +5,7 @@ import { Position } from '@pancakeswap/v3-sdk'
 import { LightGreyCard } from 'components/Card'
 import { DoubleCurrencyLogo } from 'components/Logo'
 import CurrencyLogo from 'components/Logo/CurrencyLogo'
-import { useStablecoinPrice } from 'hooks/useBUSDPrice'
+import { useStablecoinPrice } from 'hooks/useStablecoinPrice'
 import { formatTickPrice } from 'hooks/v3/utils/formatTickPrice'
 import { ReactNode, useState, useCallback } from 'react'
 import { unwrappedToken } from 'utils/wrappedCurrency'
