+++ apps/web/src/views/FixedStaking/components/StakingModalTemplate.tsx
@@ -26,7 +26,7 @@ import { useCallWithGasPrice } from 'hooks/useCallWithGasPrice'
 import { ToastDescriptionWithTx } from 'components/Toast'
 import { ApprovalState, useApproveCallback } from 'hooks/useApproveCallback'
 import { CurrencyAmount, Percent, Currency } from '@pancakeswap/sdk'
-import { useStablecoinPriceAmount } from 'hooks/useBUSDPrice'
+import { useStablecoinPriceAmount } from 'hooks/useStablecoinPrice'
 import toNumber from 'lodash/toNumber'
 import { CurrencyLogo } from 'components/Logo'
 import first from 'lodash/first'
