+++ apps/web/src/views/Ifos/components/IfoFoldableCard/IfoPoolCard/IfoCardDetails.tsx
@@ -6,7 +6,7 @@ import { BIG_ONE_HUNDRED } from '@pancakeswap/utils/bigNumber'
 import { formatNumber, getBalanceNumber } from '@pancakeswap/utils/formatBalance'
 import { DAY_IN_SECONDS } from '@pancakeswap/utils/getTimePeriods'
 import BigNumber from 'bignumber.js'
-import { useStablecoinPrice } from 'hooks/useBUSDPrice'
+import { useStablecoinPrice } from 'hooks/useStablecoinPrice'
 import { ReactNode, useMemo } from 'react'
 import { styled } from 'styled-components'
 import { multiplyPriceByAmount } from 'utils/prices'
