+++ apps/web/src/views/Ifos/components/IfoVesting/VestingPeriod/TokenInfo.tsx
@@ -6,7 +6,7 @@ import { TokenImage } from 'components/TokenImage'
 import { VestingData } from 'views/Ifos/hooks/vesting/fetchUserWalletIfoData'
 import { PoolIds } from '@pancakeswap/ifos'
 import { getBalanceNumber } from '@pancakeswap/utils/formatBalance'
-import { useStablecoinPrice } from 'hooks/useBUSDPrice'
+import { useStablecoinPrice } from 'hooks/useStablecoinPrice'
 import { multiplyPriceByAmount } from 'utils/prices'
 import { useDelayedUnmount } from '@pancakeswap/hooks'
 import Expand from './Expand'
